whole app refactor
